tests/wpt/web-platform-tests/html/semantics/embedded-content/media-elements/event_play_noautoplay.html

<!doctype html>
<html>
 <head>
  <title>{audio,video} events - play</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/common/media.js"></script>
 </head>
 <body>
  <p><a href="https://html.spec.whatwg.org/multipage/#mediaevents">spec reference</a></p>
  <audio id="a" controls>
  </audio>
  <video id="v" controls>
  </video>
  <div id="log"></div>
  <script>
test(function() {
  var t = async_test("calling play() on audio should trigger play event", {timeout:5000});
  var a = document.getElementById("a");
  a.addEventListener("play", t.step_func(function() {
    a.pause();
    t.done();
  }), false);
  a.src = getAudioURI("/media/sound_5") + "?" + new Date() + Math.random();
  a.play()
    .then(t.step_func(function() {
      assert_unreached("pause() should reject all pending play Promises");
    }))
    .catch(t.step_func(function() {
      assert_true(true, "pause() should reject all pending play Promises");
    }));
}, "audio events - play");

test(function() {
  var t = async_test("calling play() on video should trigger play event", {timeout:5000});
  var v = document.getElementById("v");
  v.addEventListener("play", t.step_func(function() {
    v.pause();
    t.done();
  }), false);
  v.src = getVideoURI("/media/movie_5") + "?" + new Date() + Math.random();
  v.play()
    .then(t.step_func(function() {
      assert_unreached("pause() should reject all pending play Promises");
    }))
    .catch(t.step_func(function() {
      assert_true(true, "pause() should reject all pending play Promises");
    }));
}, "video events - play");
  </script>
 </body>
</html>
